remove tokenizer from model
